refactor(web): декомпонувати sw.ts 643 LOC на entry + 6 шарів (initiative 0001)
Розбиваю apps/web/src/sw.ts (643 LOC) на тонкий entry + 6 sibling-
файлів у apps/web/src/sw/:

- sw/version.ts (19 LOC) — SW_VERSION + CACHE_NAMES (build-time
  defines).
- sw/cache.ts (122 LOC) — workbox precache + 2 NetworkFirst route-и
  (navigation + GET /api/*) + cacheEntryCount/listStaleCaches/
  clearAppCaches helpers.
- sw/notifiedKeys.ts (140 LOC) — IDB-persisted dedup-set
  (loadNotifiedKeys, recordNotified, pruneOldNotifiedKeys) для
  reminder-сповіщень. SW часто терплять ~30 c idle, тож без IDB
  ми б повторно вистрелили те саме сповіщення на cold-start.
- sw/reminders.ts (267 LOC) — reminder loop: routine/fizruk/
  nutrition checks, scheduleNextCheck, startReminderLoop +
  setRoutineData/Fizruk/Nutrition + getReminderState.
- sw/debug.ts (74 LOC) — buildSwSnapshot для UI «Дебаг service
  worker» + getDebugEnabled/setDebugEnabled.
- sw/messages.ts (126 LOC) — message-event dispatcher (SKIP_WAITING,
  SW_SET_DEBUG, SW_DEBUG, CLEAR_SW_CACHES, *_STATE_UPDATE,
  ROUTINE_NOTIFICATION_SENT). Виокремлено, бо це найбільш дискретна
  частина — додавання нового message type вже не вимагає
  прокручувати весь worker.

sw.ts (100 LOC) — composition root: addEventListener-и для install,
message, notificationclick, notificationclose, activate, push.
Жодної бізнес-логіки — лише делегує.

tsconfig.sw.json: include тепер 'src/sw.ts' + 'src/sw/**/*.ts',
тож SW-tsc бачить нові файли.
tsconfig.json: exclude додатково 'src/sw' — DOM-tsc не плутається
з WebWorker types.

eslint.config.js: знімаю файл з max-lines allowlist — нові файли
≤ 267 LOC, max-lines: 600 тримає себе сам.

Vite build: SW bundle 32.39 kB (idential до попереднього).
TypeScript: tsconfig.sw.json clean, tsconfig.json clean.

Phase 2 PR initiative-0001-module-decomposition.md, файл #5 з 5
(заключний).
