+++ apps/web/tsconfig.json
@@ -32,5 +32,5 @@
     }
   },
   "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist", "src/sw.ts"]
+  "exclude": ["node_modules", "dist", "src/sw.ts", "src/sw"]
 }
