+++ apps/web/tsconfig.sw.json
@@ -14,6 +14,6 @@
     "forceConsistentCasingInFileNames": true,
     "types": []
   },
-  "include": ["src/sw.ts"],
+  "include": ["src/sw.ts", "src/sw/**/*.ts"],
   "exclude": ["node_modules", "dist"]
 }
